+++ packages/frontend/src/pages/interop/intent-bridges/components/IntentBridgesTransferTrigger.tsx
@@ -1,8 +1,7 @@
-import { ProjectId } from '@l2beat/shared-pure'
 import { useQuery } from '@tanstack/react-query'
-import { type ReactNode, useMemo, useState } from 'react'
+import type { ReactNode } from 'react'
+import { InteropTransferDetailsTrigger } from '~/pages/interop/components/InteropTransferDetailsTrigger'
 import { useTRPC } from '~/trpc/React'
-import { TransferDetailsDialog } from '../../components/table/transfer-count-cell/TransferCountCell'
 import { useIntentBridgesSelectedChains } from '../utils/IntentBridgesSelectedChainsContext'
 
 export function IntentBridgesTransferTrigger({
@@ -19,12 +18,7 @@ export function IntentBridgesTransferTrigger({
   children: ReactNode
 }) {
   const trpc = useTRPC()
-  const [isOpen, setIsOpen] = useState(false)
   const { selectedChains } = useIntentBridgesSelectedChains()
-  const dialogSelection = useMemo(
-    () => selectionForApi ?? { from: selectedChains, to: selectedChains },
-    [selectionForApi, selectedChains],
-  )
   const { data } = useQuery(
     trpc.interop.intentBridges.queryOptions({
       from: selectedChains,
@@ -33,23 +27,15 @@ export function IntentBridgesTransferTrigger({
   )
 
   return (
-    <>
-      <button
-        type="button"
-        onClick={() => setIsOpen(true)}
-        className={className}
-      >
-        {children}
-      </button>
-      <TransferDetailsDialog
-        protocol={{ ...protocol, id: ProjectId(protocol.id) }}
-        type={undefined}
-        tokenId={tokenId}
-        snapshotTimestamp={data?.snapshotTimestamp}
-        selectedChains={dialogSelection}
-        isOpen={isOpen}
-        setIsOpen={setIsOpen}
-      />
-    </>
+    <InteropTransferDetailsTrigger
+      protocol={protocol}
+      tokenId={tokenId}
+      selectedChains={selectedChains}
+      selectionForApi={selectionForApi}
+      snapshotTimestamp={data?.snapshotTimestamp}
+      className={className}
+    >
+      {children}
+    </InteropTransferDetailsTrigger>
   )
 }
